From: 	Bort, Paul Subject: 	RE: Patch for Dia templates for Postgresql_AutoDoc Date: 	Tue, 22 Jun 2004 15:23:24 -0400
I've used the same method of shortening as the constraint you mentioned, but
I made it a new entry (table_comment_dia) in the hash rather than mangle the
existing one. I tried to follow the existing naming conventions and coding
style. I picked a total of 45 characters because that seems to be about the
right size on my screen. In testing I also found a bug in Dia that causes a
crash if a comment is too long, but that seems to be in the 100+ character
range, so I'm not worried about 45.

Here are the patches for postgresql_autodoc.pl, dia.tmpl, and
zigzag.dia.tmpl.